a REST web service implementation class
The @Path annotation specifies that resources made available by this
class will be accessible by the relative URI "/course".  The getCourse
method has many annotations. Let's discuss them one at a time.  The @GET
annotation specifies that when the relative URI (as specified by @Path
on the CourseService class) "/course" is called using the HTTP GET
method, then this method will be invoked.  @Produces
(MediaType.APPLICATION_JSON) specifies that this method generates a JSON
output. If the client specifies the accepted mime types, then this
annotation would be used to resolve the method to be called, if more
than one method is annotated with @GET (or, for that matter, any of the
other HTTP method annotations).  For example, if we have another method
called getCourseJSON annotated with @ GET but producing data with
different mime types (as specified by @Produces), then an appropriate
method will be selected on the basis of the mime type requested by the
client. The mime type in the @Produces annotation also tells the JAX-RS
implementation the mime type of the response when marshalling the Java
object that is returned from the method. For example, in the getCourse
method, we return an instance of Course, and the mime type specified in
@Produces tells Jersey to generate an XML representation of this
instance.  The @Path annotation can also be used at the method level to
specify subresources.  The value specified in @Path at the method level
is relative to the path value specified at the class level. The resource
(in this case, Course) with ID 20 can be accessed as /course/get/20. The
complete URL can be
http://<serveraddress>:<port>/<app-name>/course/get/10.  Parameter names
in the path value are enclosed in {}.  Path parameters need to be
identified in method arguments by using the @PathParam annotation and
the name of the parameter as its value. The JAX-RS implementation
framework matches the path parameters with arguments matching the
@PathParam annotation and appropriately passes the parameter values to
the method.