to tell the Jersey framework what packages it needs to scan to look for REST
to create a subclass of Application. However, instead of directly
subclassing Application, we will subclass the ResourceConfig class of
Jersey, which in turn extends Application.
We have used the @ApplicationPath annotation to specify the URL mapping
for the REST services implemented using JAX-RS. All @Path URIs on the
resource implementation classes will be relative to this path. For
example, the "/course" URI that we specified for the CourseService class
would be relative to "services", specified in the @ApplicationPath
annotation.