let's say we want to create a JSON response
Instead of generating an XML response, let's say we want to create a
JSON response, because it would be much easier to consume a JSON
response from JavaScript in a web page than an XML response. To add
support for creating a JSON response, we will have to change the value
of the @Produces annotation in the CourseService class. Currently, it is
set to MediaType.APPLICATION_XML and we want to set it to
MediaType.APPLICATION_JSON.