Implementing the REST POST request
Let's now implement a POST request. We will implement a method to add a
course in the CourseService class, which is our web service
implementation class in the CourseManagementREST project.  As in the
case of the getCourse method, we won't actually access the database but
will simply write a dummy method to save the data. Again, the idea is to
keep the example simple and focus only on the JAX-RS APIs and
implementation. Open CourseService.java and add following methods: